Typo in XSS prevention
I believe there's a small error in the last RC2 in the "How Do I Prevent This?" paragraph.
Ruby 3.0 doesn't exist (version 2.4 is the latest), I believe the writer wanted to refer to Ruby on Rails 3.0 (Ruby is the general language, Ruby on Rails is a specific web framework), which started to automatically escaping unsafe input. Moreover RoR 3.0 is from august 2010, so we can now simply refer to just Ruby on Rails safely I believe.
Also React JS should be React.js (old) or simply React as it's called nowadays.